Precall text result completed.
